Completed ML Project
The final project in Machine Learning encompasses and covers all the topics covered in the course. As per the instructions, the dataset used in the project is “Toronto Bicycle Theft”, which reports all the bicycle theft cases in a specific time frame, in the different areas of the city of Toronto. The cases enlists the occurrence date and time of the theft followed by the reporting date and time and the status (whether it is stolen or recovered). For using the brief data instead of using 31970 instances, the data belonging to the year 2022 is extracted and hence, the project is worked on 2923 instances of bicycle theft.